requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
